+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools", "wheel"]
+requires = ["setuptools", "wheel", "cmake", "nanobind"]
 build-backend = "setuptools.build_meta"
 
 [pytest]
